src/requirements.txt
langchain==0.3.27
langchain-openai==0.3.28
openai==1.98.0
PyGithub==2.7.0
pydantic==2.11.7
pydantic-settings==2.6.1
python-dotenv==1.2.2
pytest==8.3.3
pytest-asyncio==0.24.0
